feat(theater): TicketSeller 리팩토링
