+++ src/test/java/com/wanted/preonboarding/unit/theater/TicketSellerTest.java
@@ -1,0 +1,4 @@
+package com.wanted.preonboarding.unit.theater;
+
+public class TicketSellerTest {
+}
